tfa: add support for passing a dram config dtb blob to u-boot
Similar to rcar, dram config could be passed to u-boot using a
device-tree blob so that u-boot can inherit any memory identification
already performed.

Reading this blob before u-boot's own dram_init however requires an SPL
which is not currently used.
Therefore the feature is not enabled for now, and provided as a proof of
concept for future development.