tfa: support passing dram start address and size to u-boot
Implement a second option for passing dram information to U-Boot,
without DTB blob  by directly using arg2 and arg3.

This solution has much smaller code-size in both u-boot and tfa, and
unlike the blob the argument registers are accessible early on by u-boot.

This feature is also not yet enabled to allow further discussions and
development.